file_and_hash_multiple_files.py v0.1
The script takes filename as argument, including a wildcard (works sometimes) and creates a file of the hash in the location of the file that's being hashed.  Has ability to be recursive, taking -r argument.

The problem I have is when it encounters a folder that would be included with wildcard e.g.
searching for: test* -r
and a subfolder being /test...